added class for sound passing
moved audio test code to own class SoundPassThrough
